docs: install graphite-query as `graphite` command
Change the brew formula template (in publish.yml) and README to install
graphite-query's wrapper script as `graphite` instead of `graphite-query`.
`brew install graphite-query` still works (formula name unchanged),
but the binary on PATH is now `graphite`.

Tap repo was updated manually for v0.1.0-rc1.
