+++ graphite-query/src/main/kotlin/io/johnsonlee/graphite/cli/Main.kt
@@ -4,6 +4,6 @@ import picocli.CommandLine
 import kotlin.system.exitProcess
 
 fun main(args: Array<String>) {
-    val exitCode = CommandLine(QueryCommand()).execute(*args)
+    val exitCode = CommandLine(GraphiteCommand()).execute(*args)
     exitProcess(exitCode)
 }
